tweet_collector/requirements.txt
tweepy
psycopg2-binary
pymongo
python-decouple